Etch A Sketch Draft
Created a grid container and wrote a function to add cells to the container. Formatted the cells to fit the size of the container. Added a "hover" effect to change the color of the cells. Created a css sheet to color the cells and orient them about the grid container.